+++ DogBarberShop_service/src/main/java/cz/fi/muni/pa165/dogbarber/service/DogService.java
@@ -23,6 +23,6 @@ public interface DogService {
     
     void unsubscribeDogForAService(Dog dog, Service service);
     
-    BigDecimal getTotalPrice(Long dogId);
+    double getTotalPrice(Long dogId);
     Dog update(Dog dog);
 }
